requirements.txt
appdirs==1.4.3
async-generator==1.9
attrs==17.4.0
cassandra-driver==3.13.0
Cython==0.27.3
flake8==3.5.0
isort==4.3.3
mccabe==0.6.1
packaging==16.8
pluggy==0.6.0
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
pyparsing==2.2.0
pytest==3.4.0
pytest-asyncio==0.8.0
six==1.11.0
tox==2.9.1
virtualenv==15.1.0
